ocr_system/requirements.txt
# =============================================
# OCR System Requirements
# Deep Learning OCR untuk Validasi Bukti Transfer Bank Indonesia
# =============================================

# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# PaddleOCR & Dependencies (using installed versions)
paddlepaddle==3.2.1
paddleocr==3.3.1
shapely==2.0.2
pyclipper==1.3.0.post5
imgaug==0.4.0

# Image Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
albumentations==1.3.1

# Deep Learning
torch==2.1.1
torchvision==0.16.1
onnxruntime==1.16.3

# Database
pymysql==1.1.0
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0

# Utilities
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
python-dateutil==2.8.2
regex==2023.10.3

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Optional: GPU support (uncomment if using CUDA)
# paddlepaddle-gpu==2.5.2
